Dissolve importing.py. SemanticRefIndexSettings goes to semrefindex and ConversationSettings to convutils.py
